Cleanup tests, fix them and stop crying like a cry baby
