+++ app-test.js
@@ -1,48 +1,20 @@
-const {
-  setupExpressServer,
-  serveOctoFarmNormally,
-  ensureSystemSettingsInitiated
-} = require("./app-core");
-const { setupEnvConfig, fetchOctoFarmPort } = require("./app-env");
-const { serveDatabaseIssueFallback } = require("./app-fallbacks");
+const { setupExpressServer, serveOctoFarmRoutes } = require("./app-core");
+const { setupEnvConfig } = require("./app-env");
+const { ensureSystemSettingsInitiated } = require("./app-core");
 
-let httpListener;
-
+/**
+ * Setup the application without any async work
+ * @returns {app}
+ */
 async function setupTestApp() {
   setupEnvConfig(true);
-  const octoFarmTestServer = setupExpressServer();
-  await ensureSystemSettingsInitiated().catch();
+  await ensureSystemSettingsInitiated();
 
-  // Save listener as state for tests
-  httpListener = await serveOctoFarmNormally(
-    octoFarmTestServer,
-    fetchOctoFarmPort()
-  );
-
-  // const stringRouteMap = JSON.stringify(getEndpoints(octoFarmTestServer).map(r => r.path));
-  return octoFarmTestServer;
-}
-
-async function setupDatabaseIssueApp() {
-  setupEnvConfig(true);
-
-  const octoFarmTestServer = setupExpressServer();
-  httpListener = await serveDatabaseIssueFallback(
-    octoFarmTestServer,
-    fetchOctoFarmPort()
-  );
-
-  // const stringRouteMap = JSON.stringify(getEndpoints(octoFarmTestServer).map(r => r.path));
-  // console.log("Server db issue routes:", stringRouteMap);
-  return octoFarmTestServer;
-}
-
-function getServer() {
-  return httpListener;
+  const newServer = setupExpressServer();
+  serveOctoFarmRoutes(newServer);
+  return newServer;
 }
 
 module.exports = {
-  setupTestApp,
-  setupDatabaseIssueApp,
-  getServer
+  setupTestApp
 };
