+++ test/api/database-issues.test.js
@@ -1,33 +1,24 @@
-const request = require("supertest");
-const { setupDatabaseIssueApp, getServer } = require("../../app-test");
-const isDocker = require("is-docker");
+const supertest = require("supertest");
+const { setupExpressServer } = require("../../app-core");
+const { setupEnvConfig } = require("../../app-env");
+const { serveDatabaseIssueFallbackRoutes } = require("../../app-fallbacks");
 
-let app = null;
+let server;
 
-/**
- * Remove and close the db and server.
- */
-afterAll(async () => {
-  if (!!app) {
-    getServer().close();
-  }
-  jest.clearAllTimers();
-});
+async function setupDatabaseIssueApp() {
+  setupEnvConfig(true);
 
-async function getOrCreateDatabaseIssueApp() {
-  if (!app) {
-    app = await setupDatabaseIssueApp();
-  }
-  return app;
+  server = setupExpressServer();
+  await serveDatabaseIssueFallbackRoutes(server);
 }
 
 jest.mock("../../server_src/runners/githubClient");
 
 describe("DatabaseIssue server", () => {
   it("should return database issue page when no database is connected", async () => {
-    app = await getOrCreateDatabaseIssueApp();
+    app = await setupDatabaseIssueApp();
 
-    const res = await request(app).get("/").send();
+    const res = await supertest(server).get("/").send();
     expect(res.statusCode).toEqual(200);
     expect(res.text).toContain(
       "Docker mode:\n" +
