--- test/api/mock-data/package.json
@@ -1,3 +1,0 @@
-{
-  "name": "octofarm-test"
-}
