+++ test/db-handler.js
@@ -13,6 +13,7 @@ module.exports.connect = async () => {
 
   const mongooseOpts = {
     useNewUrlParser: true,
+    useCreateIndex: true,
     useUnifiedTopology: true
   };
