+++ test/domain-shared/mock-data/package.json
@@ -1,0 +1,3 @@
+{
+  "name": "octofarm-test"
+}
